requirements.txt
fastapi
uvicorn[standard]
python-multipart
easyocr
opencv-python-headless
gspread
oauth2client
python-dotenv
